+++ app/Filament/Resources/CommandResource/Pages/EditCommand.php
@@ -3,20 +3,15 @@
 namespace App\Filament\Resources\CommandResource\Pages;
 
 use App\Filament\Resources\CommandResource;
-use Filament\Actions;
+use App\Filament\Traits\HasEditFormActionsWithIcons;
+use App\Filament\Traits\HasEditActionsWithIcons;
 use Filament\Resources\Pages\EditRecord;
 
 class EditCommand extends EditRecord
 {
-	use CommandResource\Traits\PreparesForValidation;
+	use CommandResource\Traits\PreparesForValidation,
+		HasEditActionsWithIcons,
+		HasEditFormActionsWithIcons;
 
     protected static string $resource = CommandResource::class;
-
-	protected function getActions(): array
-    {
-        return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
-        ];
-    }
 }